2.0.2

+++ src/libmongoc/examples/example-gridfs.c
@@ -78,7 +78,7 @@ main (int argc, char *argv[])
             break;
          }
 
-         if (fwrite (iov.iov_base, 1, r, stdout) != r) {
+         if (fwrite (iov.iov_base, 1, r, stdout) != (size_t) r) {
             MONGOC_ERROR ("Failed to write to stdout. Exiting.\n");
             exit (1);
          }
